+++ pages/markdown/misc/changelog.md
@@ -7,6 +7,7 @@
 
 
 * **Fixed Password Reset Case-Sensitivity:** Updated password resets to no longer be case-sensitive when asking for email addresses.
+* **Improved Autocomplete for Assignments:** Improved autocomplete for when teachers are creating assignments.  If the assignment title contains "essential" in it, Alu will now automatically fill the tag query to be: "Unit # AND essential".
 
 
 ## Overflow Bucket - May 03 - 0.6.3